Added account data validation
I worked on one bug for 2 hours. Ive added the ability of the backend to initialize a student account while making sure that the account number isnt already in use. Ive fully tested this but I have not tested the ability of the teacher initialization to do the same. Before working on the pass creation I will test the teacher initialization with my email acting as the teacher acount